client: Support generic Bearer token auth
The client implementation does not really care about the details of the
bearer token used for authentication, for it it is simply a ByteString
that is copied to an HTTP header.

Generalise the client implementation to work with arbitrary Bearer
tokens, not only JWT. The actual code path used for JWT tokens remains
the same.